requirements.txt
pymongo~=4.4.1
python-dotenv~=1.0.0
pydantic~=2.1.1
fastapi~=0.101.1
uvicorn==0.23.2
passlib~=1.7.4
python-jose~=3.3.0
python-multipart~=0.0.6
bcrypt~=4.0.1
pandas~=2.1.1
SQLAlchemy~=2.0.22
psycopg2-binary~=2.9.9
redis~=5.0.1
haversine~=2.8.1